Analyzing employee salaries in America in the field of data analysis
Analysis of employee salaries data in the information technology sector in the data analysis sector in America. This data was taken from the Kaggle website. All analysis and data engineering operations were performed on this data using the Python programming language by refining the data, extracting some important information from it, and filling in the missing data with similar values. And get rid of outlier values ​​and perform various encryption operations on it so that the data set becomes ready for use and input into machine learning models for prediction and learning.